CLI Syntax #67
- Added UT placeholders
- Implemented parser
- Refactoring
- UT :+1:
