+++ ModelCompilerUI.UnitTest/CommandLineSyntax/UnitsOptionsParserUnitTest.cs
@@ -1,0 +1,22 @@
+﻿//__________________________________________________________________________________________________
+//
+//  Copyright (C) 2022, Mariusz Postol LODZ POLAND.
+//
+//  To be in touch join the community at GitHub: https://github.com/mpostol/OPC-UA-OOI/discussions
+//__________________________________________________________________________________________________
+
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace OOI.ModelCompilerUI.CommandLineSyntax
+{
+  [TestClass]
+  public class UnitsOptionsParserUnitTest
+  {
+    [TestMethod]
+    public void TestMethod1()
+    {
+      //TODO CLI Syntax #67
+      Assert.Inconclusive();
+    }
+  }
+}
