+++ ModelCompilerUI.UnitTest/ModelCompilerUI.UnitTest.csproj
@@ -90,6 +90,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CommandLineSyntax\DotNetStackOptionsParserUnitTest.cs" />
+    <Compile Include="CommandLineSyntax\UnitsOptionsParserUnitTest.cs" />
+    <Compile Include="CommandLineSyntax\UpdateHeadersOptionsParserUnitTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ModelCompilerAPIInternalUnitTest.cs" />
     <Compile Include="IntegrationUnitTest.cs" />
