+++ ModelCompilerUI/CommandLineSyntax/CompilerOptions.cs
@@ -75,76 +75,4 @@ namespace OOI.ModelCompilerUI.CommandLineSyntax
       public const string ReleaseCandidate = "rc";
     }
   }
-
-  [Verb("stack", false, HelpText = "Generates code for the core model (not used for vendor defined models)")]
-  public class DotNetStackOptions
-  {
-    [Option('n', OptionsNames.DotNetStackPath, HelpText = stackHelp, Group = "StackType", Required = true)]
-    public string DotNetStackPath { get; set; }
-
-    [Option('a', OptionsNames.AnsiCStackPath, HelpText = ansicHelp, Group = "StackType", Required = true)]
-    public string AnsiCStackPath { get; set; }
-
-    private const string ansicHelp = "Generates the ANSI C stack code for the core model (not used for vendor defined models). The path to use when generating ANSI C stack code.";
-    private const string stackHelp = "Generates the .NET stack code for the core model (not used for vendor defined models). The path to use when generating .NET stack code.";
-
-    private static class OptionsNames
-    {
-      public const string AnsiCStackPath = "ansic";
-      public const string DotNetStackPath = "stack";
-    }
-  }
-
-  [Verb("units", false, HelpText = "Generates the OPC UA Engineering Units CSV from the official UNECE table of units.")]
-  public class UnitsOptions
-  {
-    [Option(OptionsNames.Annex1Path, HelpText = annex1Help, Required = true)]
-    public string Annex1Path { get; set; }
-
-    [Option(OptionsNames.Annex2Path, HelpText = annex2Help, Required = true)]
-    public string Annex2Path { get; set; }
-
-    [Option(OptionsNames.UnitsOutputPath, HelpText = outputHelp, Required = true)]
-    public string OutputPath { get; set; }
-
-    private const string annex1Help = "The path to the UNECE Annex 1 CSV file.";
-    private const string annex2Help = "The path to the UNECE Annex 2/3 CSV file.";
-    private const string outputHelp = "The units output directory.";
-
-    private static class OptionsNames
-    {
-      public const string Annex1Path = "annex1";
-      public const string Annex2Path = "annex2";
-      public const string UnitsOutputPath = "output";
-    }
-  }
-
-  [Verb("update-headers", false, HelpText = "Updates all files in the output directory with the OPC Foundation MIT license header.")]
-  public class UpdateHeadersOptions
-  {
-    [Option(OptionsNames.InputPath, HelpText = inputPathHelp, Required = true)]
-    public string InputPath { get; set; }
-
-    [Option(OptionsNames.FilePattern, HelpText = filePatternHelp, Required = true)]
-    public string FilePattern { get; set; }
-
-    [Option(OptionsNames.LicenseType, HelpText = licenseTypeHelp, Required = true)]
-    public LicenseType LicenseType { get; set; }
-
-    [Option(OptionsNames.Silent, HelpText = silentHelp, Required = false)]
-    public bool Silent { get; set; }
-
-    private const string inputPathHelp = "The path folders to search for files to update.";
-    private const string filePatternHelp = "The file pattern to use when selecting files.";
-    private const string licenseTypeHelp = "The type of license (MIT | MITXML | NONE).";
-    private const string silentHelp = "Suppresses any exceptions.";
-
-    private static class OptionsNames
-    {
-      public const string InputPath = "input";
-      public const string FilePattern = "pattern";
-      public const string LicenseType = "license";
-      public const string Silent = "silent";
-    }
-  }
 }
